Switched to ristrict_sr flag in search
Because it seems pretty integral to the reddit api that a /[/r/subreddit]/search can be restricted to the sub or not, I figured tying the restrict_sr flag to an argument with a default value was the safest bet. OOP conventions say that the search should probably default to restricted though, so  I have the default set to true.